Fix tag validation and simplify codebase architecture
## Summary
This commit fixes critical tag validation bugs and simplifies the codebase by consolidating duplicate tool implementations.

## 🐛 Bug Fixes
- **Fix .env file loading**: Server now properly loads default .env files when no explicit --env-file is provided
- **Fix tag validation timing**: Tag validation now occurs BEFORE todo creation, not after
- **Fix tag policy enforcement**: THINGS_MCP_AI_CAN_CREATE_TAGS=false now properly prevents AI tag creation

## 🏗️ Architecture Changes
- **Consolidate tools**: Remove obsolete ThingsTools, rename HybridTools to ThingsTools
- **Simplify configuration**: Remove unused config options, keep only essential settings
- **Maintain hybrid approach**: Fast reads via things.py + reliable writes via AppleScript

## 🗂️ Files Changed
- Removed: src/things_mcp/hybrid_tools.py (renamed to tools.py)
- Modified: src/things_mcp/tools.py (now the main ThingsTools implementation)
- Modified: src/things_mcp/config.py (fix .env loading)
- Modified: src/things_mcp/server.py (update imports)
- Simplified: .env.example (remove unused options)

## 🧪 Testing
- Configuration loading verified
- Tag validation service initialization confirmed
- All existing functionality preserved
