+++ .env.example
@@ -1,14 +1,5 @@
-# Things 3 MCP Server Configuration Example
+# Things 3 MCP Server Configuration
 # Copy this file to .env and customize the values as needed
-
-# =============================================================================
-# SERVER CONFIGURATION
-# =============================================================================
-
-# Server identification
-THINGS_MCP_SERVER_NAME=things3-mcp-server
-THINGS_MCP_SERVER_VERSION=1.0.0
-THINGS_MCP_SERVER_DESCRIPTION=MCP server for Things 3 task management integration
 
 # =============================================================================
 # APPLESCRIPT EXECUTION
@@ -19,9 +10,6 @@ THINGS_MCP_APPLESCRIPT_TIMEOUT=30.0
 
 # Number of retries for failed operations (0-10)
 THINGS_MCP_APPLESCRIPT_RETRY_COUNT=3
-
-# Execution method: url_scheme, applescript, or hybrid
-THINGS_MCP_PREFERRED_EXECUTION_METHOD=hybrid
 
 # =============================================================================
 # TAG CREATION CONTROL
@@ -40,6 +28,9 @@ THINGS_MCP_TAG_VALIDATION_CASE_SENSITIVE=false
 # Prevents accidental creation of too many tags in bulk operations
 THINGS_MCP_MAX_AUTO_CREATED_TAGS_PER_OPERATION=10
 
+# Maximum number of tags per item (1-50)
+THINGS_MCP_MAX_TAGS_PER_ITEM=20
+
 # =============================================================================
 # LOGGING
 # =============================================================================
@@ -50,47 +41,4 @@ THINGS_MCP_LOG_LEVEL=INFO
 # Optional: Log to file instead of console
 # Uncomment and set path to enable file logging
 # THINGS_MCP_LOG_FILE_PATH=/var/log/things-mcp.log
-
-# =============================================================================
-# VALIDATION
-# =============================================================================
-
-# Maximum length for todo/project titles (50-1000)
-THINGS_MCP_MAX_TITLE_LENGTH=500
-
-# Maximum length for notes (100-50000)
-THINGS_MCP_MAX_NOTES_LENGTH=10000
-
-# Maximum number of tags per item (1-50)
-THINGS_MCP_MAX_TAGS_PER_ITEM=20
-
-# Maximum number of checklist items (1-200)
-THINGS_MCP_MAX_CHECKLIST_ITEMS=50
-
-# =============================================================================
-# PERFORMANCE
-# =============================================================================
-
-# Default maximum search results (10-1000)
-THINGS_MCP_SEARCH_RESULTS_LIMIT=100
-
-# =============================================================================
-# SECURITY
-# =============================================================================
-
-# Enable input sanitization for AppleScript
-THINGS_MCP_ENABLE_INPUT_SANITIZATION=true
-
-# =============================================================================
-# THINGS 3 INTEGRATION
-# =============================================================================
-
-# Name of the Things 3 application
-THINGS_MCP_THINGS_APP_NAME=Things3
-
-# Base URL for Things 3 URL schemes
-THINGS_MCP_THINGS_URL_SCHEME_BASE=things:///
-
-# Verify Things 3 is available on startup
-THINGS_MCP_VERIFY_THINGS_AVAILABILITY=true
